requirements.txt
python-telegram-bot
Pillow
python-dotenv
rarfile
